// packages/shared/src/enums/complaint-status.enum.ts
export enum ComplaintStatus {
  SUBMITTED = 'SUBMITTED',
  TRIAGE = 'TRIAGE',
  ASSIGNED = 'ASSIGNED',
  IN_INVESTIGATION = 'IN_INVESTIGATION',
  DRAFT_RESPONSE = 'DRAFT_RESPONSE',
  QA_LEGAL_REVIEW = 'QA_LEGAL_REVIEW',
  RESPONSE_ISSUED = 'RESPONSE_ISSUED',
  AWAITING_FEEDBACK = 'AWAITING_FEEDBACK',
  CLOSED = 'CLOSED',
  APPEAL = 'APPEAL',
}